User connection + get all endpoint

+++ QuizManagement/src/main/java/com/example/quizmanagement/service/QuizManagementService.java
@@ -5,18 +5,22 @@ import com.example.quizmanagement.dto.request.QuizRequest;
 import com.example.quizmanagement.dto.response.QuestionResponse;
 import com.example.quizmanagement.dto.response.QuizResponse;
 import com.example.quizmanagement.dto.response.QuizResponseWithAnswers;
+import com.example.quizmanagement.dto.response.SimpleQuizResponse;
 import com.example.quizmanagement.exceptions.BadRequestException;
+import com.example.quizmanagement.jwt.UserDetailsImpl;
 import com.example.quizmanagement.mapper.QuizManagementMapper;
 import com.example.quizmanagement.model.quiz.Question;
 import com.example.quizmanagement.model.quiz.Quiz;
 import com.example.quizmanagement.repository.QuizRepository;
 import lombok.RequiredArgsConstructor;
 import org.springframework.context.MessageSource;
+import org.springframework.security.core.context.SecurityContextHolder;
 import org.springframework.stereotype.Service;
 import org.springframework.transaction.annotation.Transactional;
 
 import java.util.List;
 import java.util.Locale;
+import java.util.Objects;
 
 @Service
 @RequiredArgsConstructor
@@ -26,12 +30,24 @@ public class QuizManagementService {
     private final MessageSource messageSource;
 
     @Transactional
-    public QuizResponseWithAnswers save(QuizRequest quiz) {
-        Quiz entity = quizRepository.save(quizManagementMapper.toEntity(quiz));
+    public QuizResponseWithAnswers save(QuizRequest request) {
+        UserDetailsImpl loggedUser = (UserDetailsImpl) SecurityContextHolder.getContext()
+                .getAuthentication().getPrincipal();
 
+        if (!loggedUser.isCompany()) {
+            throw new BadRequestException(messageSource.getMessage("USER_IS_NOT_VALID", null,
+                    Locale.getDefault()));
+        }
+
+        Quiz entity = quizManagementMapper.toEntity(request);
+
+        entity.setCreatorId(loggedUser.getId());
+        entity.setTime(request.time());
         entity.getQuestions().forEach(q -> q.setQuiz(entity));
 
-        return quizManagementMapper.toResponseWithAnswers(entity);
+        Quiz savedEntity = quizRepository.save(entity);
+
+        return quizManagementMapper.toResponseWithAnswers(savedEntity);
     }
 
     @Transactional
@@ -59,8 +75,17 @@ public class QuizManagementService {
 
     @Transactional
     public QuizResponseWithAnswers editQuiz(QuizRequest request, long id) {
+
+        UserDetailsImpl loggedUser = (UserDetailsImpl) SecurityContextHolder.getContext()
+                .getAuthentication().getPrincipal();
+
         Quiz quiz = quizRepository.findById(id).orElseThrow(() -> new BadRequestException(messageSource
                 .getMessage("QUIZ_NOT_FOUND", null, Locale.getDefault())));
+
+        if (!Objects.equals(quiz.getCreatorId(), loggedUser.getId())) {
+            throw new BadRequestException(messageSource.getMessage("USER_IS_NOT_VALID", null,
+                    Locale.getDefault()));
+        }
 
         quiz.setTechnology(request.technology());
         quiz.setTitle(request.title());
@@ -72,6 +97,21 @@ public class QuizManagementService {
     public void deleteQuiz(long id) {
         Quiz quiz = quizRepository.findById(id).orElseThrow(() -> new BadRequestException(messageSource
                 .getMessage("QUIZ_NOT_FOUND", null, Locale.getDefault())));
+        UserDetailsImpl loggedUser = (UserDetailsImpl) SecurityContextHolder.getContext()
+                .getAuthentication().getPrincipal();
+        if (!Objects.equals(quiz.getCreatorId(), loggedUser.getId())) {
+            throw new BadRequestException(messageSource.getMessage("USER_IS_NOT_VALID", null,
+                    Locale.getDefault()));
+        }
         quizRepository.delete(quiz);
+    }
+
+    public List<SimpleQuizResponse> getAllQuizByCompany() {
+        UserDetailsImpl loggedUser = (UserDetailsImpl) SecurityContextHolder.getContext()
+                .getAuthentication().getPrincipal();
+
+        List<Quiz> allByCreatorId = quizRepository.findAllByCreatorId(loggedUser.getId());
+
+        return allByCreatorId.stream().map(quizManagementMapper::toSimpleResponse).toList();
     }
 }
